Update Pypeline.py
Last Modification: Cleaned up TaskOpsUtils._run_opperation() a bit, though it could probably improve further by a name change.
Added a generator object as an attribute to Pipeline (Pipeline.tasks) which will provide a generator of nested dicts {pipe_name: {task_name: task}}
    I plan to use this feature in a future update for multithreading.
Added __getattribute__ and __getitem__ to Pipe and Pipeline to allow dot-notation navigation between objects.

Note: Next should be restricting setattribute, and allowing setitem to adding pipes/tasks to a pipeline/pipe, respectively.
There should also be some functionality to scope a wrangler/loader to a task, pipe, or entire pipeline--while still letting set wrangler/loaders at more granular levels override the parent counterpart.

Note: I also created a file pipelineutils which contains a credential manager (for PyKeePass) and a Pipeline-Scopable wrangler that maps input [dict,] commands to Pandas operations. This would idealy be used as a convenient way to store Pandas operations as JSON, along with the JSON representation of the Task it applies to. The mapping functionality still requires some debugging, but in theory it will allow regular methods, loc, operators (+, -, =, ...), and the ability to store named _variables for later use. The mapping verifies that passed arguments are attributes of a DataFrame, and uses `getattr()` for security reasons.